WIP: Adapted prediction function for runtime injection.
The intermediate values for the selected layer are no longer saved due
to out of memory issues. Instead, the network is always executed from
the beginning and the batch is injected during runtime. While fixing the
memory issue, this will inevitabley lead to slower performance as the
network has to be executed from the beginning every time.